size1/requirements.txt
# DO NOT include azure-functions-worker in this file
# The Python Worker is managed by Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions
azure-functions-durable
pandas
numpy
scikit-learn
-f https://download.pytorch.org/whl/torch_stable.html
torch==2.0.1+cpu 
torchvision==0.15.2+cpu  
torchmetrics
